third_party/WebKit/LayoutTests/external/wpt/safe-html-types/innerHTML.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.js"></script>
<body>
<script>
  promise_test(t => {
    return SafeHTML.escape(unescapedHTML).then(html => {
      var d = document.createElement('div');
      d.innerHTML = html;
      assert_equals(d.innerText, unescapedHTML);
    });
  }, "innerHTML = SafeHTML.escape().");

  promise_test(t => {
    return SafeHTML.unsafelyCreate(unescapedHTML).then(html => {
      var d = document.createElement('div');
      d.innerHTML = html;
      assert_equals(d.innerText, unescapedText);
    });
  }, "innerHTML = SafeHTML.unsafelyCreate().");
</script>
